// src/main/MusicDownloaderAPi/types/DownloadMetaData_Type.d.ts
export interface Format {
  format_id: string
  format_note?: string
  ext: string
  protocol: string
  acodec: string
  vcodec: string
  url: string
  width: number | null
  height: number | null
  fps: number | null
  rows?: number
  columns?: number
  fragments?: Fragment[]
  audio_ext: string
  video_ext: string
  vbr: number | null
  abr: number | null
  tbr: number | null
  resolution: string
  aspect_ratio: number | null
  filesize_approx: number | null
  http_headers: HttpHeaders
  format: string
  asr?: number | null
  filesize?: number | null
  source_preference?: number
  audio_channels?: number | null
  quality?: number
  has_drm?: boolean
  language?: string | null
  language_preference?: number
  preference?: number | null
  container?: string
  available_at?: number
  downloader_options?: DownloaderOptions
  manifest_url?: string
}

export interface Fragment {
  url: string
  duration: number
}
// ...
export interface DownloaderOptions {
  http_chunk_size: number
}

export interface HttpHeaders {
  'User-Agent': string
  Accept: string
  'Accept-Language': string
  'Sec-Fetch-Mode': string
}